feat: Move language handling functions to a new utility module for better organization
