+++ AbbyBot-API/app/routes/server_settings.py
@@ -1,41 +1,7 @@
 from flask import Blueprint, jsonify, request
-from ..utils.db import get_db_connection
+from ..utils.server_settings_utils import get_current_language, update_language
 
 server_settings_bp = Blueprint('server_settings', __name__)
-
-def get_current_language(guild_id):
-    conn = get_db_connection("AbbyBot_Rei")
-    cursor = conn.cursor()
-    
-    try:
-        query = "SELECT guild_language FROM server_settings WHERE guild_id = %s;"
-        cursor.execute(query, (guild_id,))
-        result = cursor.fetchone()
-        return result[0] if result else None 
-    finally:
-        cursor.close()
-        conn.close()
-
-def update_language(guild_id, guild_language):
-    conn = get_db_connection("AbbyBot_Rei")
-    cursor = conn.cursor()
-
-    try:
-        current_language = get_current_language(guild_id)
-        if current_language == guild_language:
-            return -1  
-
-        query = """
-            UPDATE server_settings 
-            SET guild_language = %s 
-            WHERE guild_id = %s;
-        """
-        cursor.execute(query, (guild_language, guild_id))  
-        conn.commit()  
-        return cursor.rowcount  
-    finally:
-        cursor.close()
-        conn.close()
 
 @server_settings_bp.route('/update-language', methods=['POST'])
 def update_guild_language_route():
